add PM departemnt
